sliced up daily tasks into a bunch of functions in a separate file
